Adds a Flip Period Knob to the Stereo Flipper
This is a pretty basic update which allows the user to choose how long each flip period is within a generic editor interface. The concept of the Audio Plugin is proven at this point and I would say it creates an interesting stereo effect which is also an original effect. Up next, maybe more functionality and a better user interface along with sync options for actual note lengths per flip.